requirements.txt
flask>=2.0.0
google-generativeai>=0.3.0
Pillow>=9.0.0
werkzeug>=2.0.0
python-dotenv>=0.19.0